added user attachment section

+++ app/public/js/main.js
@@ -1,5 +1,4 @@
 clientsList = document.getElementById('clients-list');
-
 if (clientsList) {
     clientsList.addEventListener('click', e => {
         if (e.target.className === 'btn btn-link delete-client') {
@@ -22,6 +21,21 @@ if (clientInsObjects) {
                 let id = e.target.dataset.id;
 
                 fetch(`/client/insurance/delete/${id}`, {
+                    method: 'DELETE'
+                }).then(res => window.location.reload())
+            }
+        }
+    });
+}
+
+attachmentList = document.getElementById('attachment-list');
+if (attachmentList) {
+    attachmentList.addEventListener('click', e => {
+        if (e.target.className === 'badge badge-danger delete-attachment') {
+            if (confirm('Are you sure?')) {
+                let id = e.target.dataset.id;
+
+                fetch(`/client/delete-attachment/${id}`, {
                     method: 'DELETE'
                 }).then(res => window.location.reload())
             }
